+++ resources/views/client/create.blade.php
@@ -14,7 +14,7 @@
 
 			<form action="{{ route('client.store') }}" method="post">
 
-				<input type="hidden" name="user_id" value="{{ Auth::user()->id }}">
+				<input type="hidden" name="user_id" value="{{ Auth::user()->id }}" />
 
 				{!! csrf_field() !!}