+++ change-log.md
@@ -1,7 +1,10 @@
 ### Change Log
 
+### 4.9.16 - 1.1.0
+- Set up grunt browserify
+- React components for create client button, create client form
 
-### 4.9.16 - 1.0.0
+### 4.9.16 - 1.1.0
 - Client index view
 - Client create view
 - Client store method